alpha.02
First working version.  Two known bugs:
(1) the entrance event to the cave doesn't get randomized, it always drops you off in the normal place.  This is because that event happens after a branch choice that will need to be updated during.
(2) When you escape Umaro's cave, the character sprite isn't drawn on the map for some reason, and the music doesn't change like it's supposed to.  The sprite comes back when you walk into another room, but the music stays the same.