alpha.03
Two bugfixes:
(1) Added an event patching routine for special cases. Use this approach to fix the entrance event to the cave, which is now randomized correctly. However it has a graphical glitch and doesn't start the music.  Those are post-map-transition parts of the vanilla event that will need to be patched in.
(2) Added "hidden character" bit to event_exit_info; and added checking these bits to patch in "show character" after map transition when appropriate.  Fixed.