alpha.22
Updates to Doors.map_oneways():
- Added a check to peek ahead and see if connecting to a zone will cause errors due to forced or shared connections (lines 606 - 660)
- Separated handling of true forced exits and shared exits:
-- Created new dictionary rooms.shared_oneways[exit_ID] = [list of shared exit_IDs] for cases where two one-way exits should connect to the same destination.
-- rooms.forced_connections[exitID] = [list of entrance_IDs] is now just for truly forced connections, i.e. exitID MUST be connected to one of [list of entrance_IDs].  Technically this could include doors but is currently just one-ways connecting virtual rooms (e.g. esper mountain room2 jumps).
-- added new dictionary, doors.sharing[d] = [d_shared], parallel to doors.forcing[d] = [d_forced].
- changed to random selection of forced (or shared) exits from available pool.
- Made walk check phase more robust: code now verifies that there are available downstream exits from the connection AND upstream entrances from the connector.  The upstream entrances was previously handled implicitly by the loop mechanism (I think), but there may have been edge cases and I think this is less ambiguous.
- Added lots of state description triggered by self.verbose = True

Note that this still fails a lot, although not as much as before.  The errors still seem to appear due to forced connections coming back into the walk and creating a state with no downstream exits.

I have a concept for fixing this problem: connect all the forced exits at the beginning to create multiple walks, then follow similar rules for walking, with added checks to make sure closed walks aren't created if there are more than one.  This will get all the forced exits out of the way at the beginning (and shared exits?); it's the oneways equivalent of mapping all the dead-end zones first, hopefully it will simplify the code too.