alpha.24
Feature update:
- Changed previous behavior of "-dra" to --door-randomize-dungeon-crawl "-drdc"
- Added new --door-randomize-all "-dra"

The previous -dra would always combine all of the randomized doors into a single, massive dungeon that you could enter from multiple places.  While a cool mode, this isn't exactly what we expect from --door-randomize-all, which is that the dungeons may or may not be connected, and e.g. you could walk into Esper Mtn and just find a single dead-end treasure room.  I renamed the old format to "Dungeon Crawl" -drdc, and added a new format "All" -dra to capture the expected "all" behavior.

Also added the capability to handle virtual doors (doors that are necessary for the logic of randomization, but are not represented by real doors in the game), these by convention come in pairs with values 8###--9### (e.g. 8001, 9001) and are always forced to connect.