alpha.25
Added franklin's updated warp code (sends player to WoB Narshe) & made all maps warpable in the case of -door-randomize-all (-dra).  Also gives the player 99 warp stones in the case of -debug.

This is a provisional solution to the one-way-loops problem for -dra.  Arguably this solves the problem, but TBD on whether it's how we actually want the game to play.