alpha.30
Bookkeeping & a few bugfixes for magitek factory
- fixed error in event 2028 transition address
- added event descriptions to map_exit_extra
- make NPC disappear after beating chad in Owzer's Mansion

The Cid elevator ride and minecart start event is going to be a problem.  maps.mod() currently executes before events.mod(), but the edit to the Cid event happens in the latter.  Currently this means that we get the conversation in the elevator and in the basement.  We could
(1) duplicate the edits in maps.mod(),
(2) somehow tell events.magitek_factory() the new address it should modify,
(3) try running maps.mod() after events.mod() [but this could cause different conflicts]
(3) something more clever?  Like read in all affected event scripts to memory, do all modifications in memory, then write all modifications & pointers to the rom afterward?  This might let us "free" that memory to be overwritten too.