alpha.38
Bugfixes for parse.simple_parser(); tested delete_nops with -dra and -drdc and it works, so turned it back on.

Opcodes with variable parameters are almost fully implemented:
- 0x00 -- 0x35 (Action queues, terminated by 0xff)
- 0x73 and 0x74 (replace background)
- 0xb0 (repeating block of commands, terminated by 0xb1 or 0xbc [+2])
- 0xb6 (branch after dialog box choices)
There is a known remaining issue with 0xb6: there's no way to tell locally in the event code how many dialog options there were, and so no way to know how many 3-byte addresses to read here.  Most dialog choices are 2 options, so for now I hardcoded that in and hope any entrance events don't have more.