beta.04: improved logic with keys & forced connections
There was a major issue when putting a forced connection inside a lock.  (This is needed in Daryl's tomb to accurately represent the turtle #2 room.)
METHOD:
- Put forced pits into node.locked, without a findable key, e.g. {'forced': [3060]}
- Only remove pits from node.locked when they are forced.
- Don't assess locked pits in network.is_attachable(node)
- When locked forced connections are unlocked, connect them with highest priority.

This update goes a long way toward fixing it; but there's still a bad interaction between locked forced connections that can cause -dra or -drdc to take forever.  I'm pushing the update for posterity, but leaving -drdt out of -dra for now.