beta.05: walk logic for locked forced connections
This includes several updates to improve walk logic in the case when forced connections are behind locks (such as Daryl's Tomb turtle room 2).  In these cases, the forced connections should be counted in network validity analysis, even though they haven't technically been connected yet.

Runs with -dra or -drdc including Daryl's Tomb now almost always compile, so we are going with it.

Also includes bugfixes in network_connect() (actions should always happen on net_state, not self).

Some code commenting in doors to see if it breaks anything; if not, we'll delete the comments in a future push.