beta.03b:  -drpt phantom train complete, untested
Code compiles.  Haven't tested it yet.  Had to include lots of metadata for the event-tiles-as-doors in data/map_exit_extra.py.

Includes a bugfix in attach_dead_ends() to prevent connecting a dead end to a room for which the only other exit is locked by a key in that room.  (This was occasionally causing uncompleteable networks in -drpt.)